+++ app/logging/logger.py
@@ -76,6 +76,8 @@ def configure_logging(
     # Prevent logs from standard libraries from drowning out app logs
     logging.getLogger("uvicorn.access").setLevel(logging.WARNING)
     logging.getLogger("uvicorn.error").setLevel(logging.WARNING)
+    logging.getLogger("httpx").setLevel(logging.WARNING)
+    logging.getLogger("huggingface_hub").setLevel(logging.WARNING)
 
 
 configure_logging()